Adapt address functions to 64 bit

+++ Common/Helper.Memory.h
@@ -3,9 +3,10 @@
 #include <string>
 #include <span>
 #include <functional>
+#include <cstdint>
 
 namespace common::helper::memory {
-    DWORD ResolveAddress(std::span<const DWORD> offsets, bool doNotValidateLastAddress);
+    uintptr_t ResolveAddress(std::span<const DWORD> offsets, bool doNotValidateLastAddress);
     std::string GetAddressConfigAsString();
     bool IsBadReadMem(const void* ptr, size_t size);
 }
